삭제 확인 confirm
나중에 직접 모달창 만들어서 적용하기!

+++ src/components/Lists/TopicListItem.js
@@ -38,6 +38,9 @@ const TopicListItem = ({ topic }) => {
       });
   };
   const onDeleteBtnCLick = () => {
+    if (!window.confirm("삭제할꺼?")) {
+      return;
+    }
     deleteTopic()
       .then(() => {
         const newTopics = topics.filter((item) => item.id !== topic.id);
